+++ loky/backend/process.py
@@ -16,7 +16,7 @@ class LokyProcess(BaseProcess):
 
     def __init__(self, group=None, target=None, name=None, args=(),
                  kwargs={}, daemon=None, init_main_module=False,
-                 env={}):
+                 env=None):
         if sys.version_info < (3, 3):
             super(LokyProcess, self).__init__(
                 group=group, target=target, name=name, args=args,
@@ -26,7 +26,7 @@ class LokyProcess(BaseProcess):
             super(LokyProcess, self).__init__(
                 group=group, target=target, name=name, args=args,
                 kwargs=kwargs, daemon=daemon)
-        self.env = env
+        self.env = {} if env is None else env
         self.authkey = self.authkey
         self.init_main_module = init_main_module